Revert "Changed salary from decimal to int in the employeeSchema.sql"
This reverts commit 090b4ca153c4469dfea1f55949269394ad357be9.

+++ employeeSchema.sql
@@ -15,7 +15,7 @@ CREATE TABLE employee(
 CREATE TABLE role(
     id INT NOT NULL AUTO_INCREMENT;
     title VARCHAR(30) NOT NULL;
-    salary INT NOT NULL;
+    salary DECIMAL(10,4) NOT NULL;
     department_id INT NOT NULL;
     PRIMARY KEY (id);
 );
